adding options for server admins to scale hollow damage in pvp battles (#2820)
* adding options for server admins to scale hollow damage in pvp battles

* adding to shields

+++ Source/ACE.Server/WorldObjects/Player_Properties.cs
@@ -1100,7 +1100,7 @@ namespace ACE.Server.WorldObjects
         /// <summary>
         /// Returns player's augmentation resistance for damage type
         /// </summary>
-        public float GetAugmentationResistance(DamageType damageType)
+        public int GetAugmentationResistance(DamageType damageType)
         {
             switch (damageType)
             {
